NumberFormatException fix for Lokalise integration data types

+++ src/main/kotlin/data/Key.kt
@@ -4,11 +4,9 @@ import com.google.gson.annotations.SerializedName
 
 data class Key(
     @SerializedName("key_id")
-    val keyId: Int,
+    val keyId: Long,
     @SerializedName("key_name")
     val keyName: KeyName,
     @SerializedName("platforms")
     val platforms: List<String>,
-//    @SerializedName("translations")
-//    val translations: List<Translation>,
 )
